// src/modules/tmdb/types/tmdb.ts
export type ContentType = 'movie' | 'tv';

export enum Trending {
  ALL = 'ALL',
  POPULAR = 'POPULAR',
  TOP_RATED = 'TOP_RATED',
  THEATHER = 'THEATHER',
}